Hack Design's Copy
I attempted to copy design of popular Hack Design's design. This is not
exact design. This one is also not responsive. It lacks lots of stuff
that Hack Design doesn't. I did rewrote whole code, none was copied.
Intent of imitation was to just polish my skills. Nothing serious. If I
did something wrong, contact me and I'll cooperate fully :)

Caution: It might be copyright violation to use this theme, since it's a
imitated one.